Fix naming and implement true GRPO LoRA training
- Rename grpo-lora-simple.py to sft-lora-simple.py (correct naming)
- Create grpo-lora-train.py: True GRPO training with LoRA adapters
- Maintain original grpo-train.py reward function (unique character counting)
- Successfully tested GRPO training with 540K trainable parameters
- Update test scripts and README documentation for clarity
- Clear distinction between GRPO, SFT, and agent training approaches

Verified functionality:
- GRPO LoRA: 20 steps, 144s, reward improvement 0.373→0.385
- SFT LoRA: 20 steps, 31s, stable supervised fine-tuning
- Agent LoRA: Chain-of-thought reasoning with custom tasks
- All save 17.2MB LoRA adapters for efficient deployment
