+++ grpo_lora_train_output/checkpoint-20/trainer_state.json
@@ -1,0 +1,134 @@
+{
+  "best_global_step": null,
+  "best_metric": null,
+  "best_model_checkpoint": null,
+  "epoch": 0.4,
+  "eval_steps": 500,
+  "global_step": 20,
+  "is_hyper_param_search": false,
+  "is_local_process_zero": true,
+  "is_world_process_zero": true,
+  "log_history": [
+    {
+      "clip_ratio/high_max": 0.0,
+      "clip_ratio/high_mean": 0.0,
+      "clip_ratio/low_mean": 0.0,
+      "clip_ratio/low_min": 0.0,
+      "clip_ratio/region_mean": 0.0,
+      "completions/clipped_ratio": 1.0,
+      "completions/max_length": 128.0,
+      "completions/max_terminated_length": 0.0,
+      "completions/mean_length": 128.0,
+      "completions/mean_terminated_length": 0.0,
+      "completions/min_length": 128.0,
+      "completions/min_terminated_length": 0.0,
+      "epoch": 0.1,
+      "grad_norm": 0.7247301340103149,
+      "kl": 1.5166588127613068e-06,
+      "learning_rate": 4.4444444444444447e-05,
+      "loss": -0.0,
+      "num_tokens": 3710.0,
+      "reward": 0.37299999594688416,
+      "reward_std": 0.04101219214498997,
+      "rewards/reward_num_unique_chars/mean": 0.37299999594688416,
+      "rewards/reward_num_unique_chars/std": 0.04101218990981579,
+      "step": 5
+    },
+    {
+      "clip_ratio/high_max": 0.0,
+      "clip_ratio/high_mean": 0.0,
+      "clip_ratio/low_mean": 0.0,
+      "clip_ratio/low_min": 0.0,
+      "clip_ratio/region_mean": 0.0,
+      "completions/clipped_ratio": 1.0,
+      "completions/max_length": 128.0,
+      "completions/max_terminated_length": 0.0,
+      "completions/mean_length": 128.0,
+      "completions/mean_terminated_length": 0.0,
+      "completions/min_length": 128.0,
+      "completions/min_terminated_length": 0.0,
+      "epoch": 0.2,
+      "grad_norm": 0.5339120626449585,
+      "kl": 1.8539559096097947e-05,
+      "learning_rate": 3.055555555555556e-05,
+      "loss": 0.0,
+      "num_tokens": 7478.0,
+      "reward": 0.3730000019073486,
+      "reward_std": 0.026870058104395866,
+      "rewards/reward_num_unique_chars/mean": 0.3730000019073486,
+      "rewards/reward_num_unique_chars/std": 0.02687005652114749,
+      "step": 10
+    },
+    {
+      "clip_ratio/high_max": 0.0,
+      "clip_ratio/high_mean": 0.0,
+      "clip_ratio/low_mean": 0.0,
+      "clip_ratio/low_min": 0.0,
+      "clip_ratio/region_mean": 0.0,
+      "completions/clipped_ratio": 0.9,
+      "completions/max_length": 128.0,
+      "completions/max_terminated_length": 22.0,
+      "completions/mean_length": 126.2,
+      "completions/mean_terminated_length": 22.0,
+      "completions/min_length": 124.4,
+      "completions/min_terminated_length": 22.0,
+      "epoch": 0.3,
+      "grad_norm": 0.0006342967390082777,
+      "kl": 3.582839271984994e-05,
+      "learning_rate": 1.6666666666666667e-05,
+      "loss": 0.0,
+      "num_tokens": 11120.0,
+      "reward": 0.3969999969005585,
+      "reward_std": 0.026870058104395866,
+      "rewards/reward_num_unique_chars/mean": 0.3969999969005585,
+      "rewards/reward_num_unique_chars/std": 0.026870055496692656,
+      "step": 15
+    },
+    {
+      "clip_ratio/high_max": 0.0,
+      "clip_ratio/high_mean": 0.0,
+      "clip_ratio/low_mean": 0.0,
+      "clip_ratio/low_min": 0.0,
+      "clip_ratio/region_mean": 0.0,
+      "completions/clipped_ratio": 1.0,
+      "completions/max_length": 128.0,
+      "completions/max_terminated_length": 0.0,
+      "completions/mean_length": 128.0,
+      "completions/mean_terminated_length": 0.0,
+      "completions/min_length": 128.0,
+      "completions/min_terminated_length": 0.0,
+      "epoch": 0.4,
+      "grad_norm": 0.5616071224212646,
+      "kl": 6.290534511208535e-05,
+      "learning_rate": 2.777777777777778e-06,
+      "loss": 0.0,
+      "num_tokens": 14826.0,
+      "reward": 0.3850000023841858,
+      "reward_std": 0.05798275358974934,
+      "rewards/reward_num_unique_chars/mean": 0.3850000023841858,
+      "rewards/reward_num_unique_chars/std": 0.05798274902626872,
+      "step": 20
+    }
+  ],
+  "logging_steps": 5,
+  "max_steps": 20,
+  "num_input_tokens_seen": 14826,
+  "num_train_epochs": 1,
+  "save_steps": 25,
+  "stateful_callbacks": {
+    "TrainerControl": {
+      "args": {
+        "should_epoch_stop": false,
+        "should_evaluate": false,
+        "should_log": false,
+        "should_save": true,
+        "should_training_stop": true
+      },
+      "attributes": {}
+    }
+  },
+  "total_flos": 0.0,
+  "train_batch_size": 1,
+  "trial_name": null,
+  "trial_params": null
+}
